fix: sidebar 레이아웃 수정
idea의 layout파일로 옮김
